Added text for first page of wizard
This text taken from grc.com. I seem to remember that all of the SQRL UI
text found there was released to the public domain, and thus able to be
included in my project, but I can't find the source for that now. If
this is not the case, please notify me, and I will remove it.